[BE] refactor: LogProperty 클래스 패키지 구조 정리 및 이동 #650
- LogProperty 관련 클래스들을 특정 도메인 패키지로 이동
